Halfway to the halfway
The Option<T> rewrite of this is underway.
It's super broken right now but I'm going to take a little break.
I had no idea that interfaces could extend other interfaces.
The keywording is a little weird but I guess it makes sense.
I also need to fix something in opre before I keep going because
throwing inside lambdas doesn't seem to work. I'll have to check on
that later.